Update controllers default response to be HTML
Adds a before action to `ApplicationController` to set the default
response to HTML for all actions.

We obviously have actions which response to JSON, CSV, ZIP, PNG, etc so
for these actions, in their own controller, we call `skip_before_action`
to bypass the default.

This prevents `ActionView::MissingTemplate` exceptions when JSON is
requested for a route which we have explicitly said responds to it.
Instead `ActionController::UnknownFormat` error will be raised which we
already catch within `ApplicationController` and return a 406 response
to the user.

This means we are now responding with 406 bad request, where we used to
respond with either a 403 forbidden or a 404 not found.

It means we can't handle invalid format, for `GeneralController#search`
we used to gracefully degrade to HTML for any unspecified formats. With
this change we return a 406 which seems like an acceptable behaviour.

We can also simplify controller `respond_to` blocks if `format.html` is
the only format specified.

Fixes #6408
Fixes #4228