Count public requests when showing existing message
When a user is making a request we show a link to existing requests made
by the body to try to prevent duplicates.

The current code includes private requests in that count, so a user
would click the link to find zero requests!

This commit fixes this by counting _public_ requests. Also improves
readability by using `#any?` instead of an integer comparison.

Fixes https://github.com/mysociety/alaveteli/issues/4220.