Fix invalid byte sequence error
Rescue from `Encoding::InvalidByteSequenceError` same as how we've
recently handled `Encoding::ConverterNotFoundError` [1].

This means for some attachments with the incorrect charset we might
scrub invalid characters. This seems to happen for right single
quotation marks, EG so "don’t" will be displayed as "dont".

This is the same behaviour as other email clients, such as Apple Mail.

[1] https://github.com/mysociety/alaveteli/pull/6451

Fixes #6046