Fix uncaught duplicate requests
Trim the leading and trailing whitespace from the request title when
checking for existing requests.

When a user is directed to signin before a request is made it results in
a `PostRedirect` being created and an email is sent to the requester to
confirm their email address. This mail includes a link to `/new` which
will finish creating their request.

The `PostRedirect` includes the all form data as it was entered into the
new `InfoRequest` form (including any extra whitespace). The comparison
to existing records can then break. We already handle whitespace in the
body to account for line endings difference but the title was a direct
comparison which would fail. Existing records have whitespace stripped
when saving into the database.

Fixes #4160