Fix locale loading when application boots
In development mode Rails might output some deprecation warnings which
don't appear in test or production. It seems Rails 6 prepares these
warnings ahead of time and they rely on the `#to_sentence` this in turn
loads the locales and but because our gettext based strings are loaded
in an initializer the available locales aren't yet set. Resulting in a
error being raised with our i18n fixes.

This change moves all the gettext and locale setup into a Railtie and
sets everything up in a `before_configuration` block which runs before
any initializers are run.

Fixes:
```
./lib/i18n_fixes.rb:54:in `available_locales': undefined method `map' for nil:NilClass (NoMethodError)
```