Update user authentication to check login token
Ensure the session contains the current login token, if not then the
user will be logged out. This prevents other sessions from remaining to
be logged in after the profile has been updated in a different browser.

This means we need to re-signin the user when the user email or password
has been changed.