Migrate RawEmail to use ActiveStorage
When the `ActiveStorage::Blob` file isn't attached then use the old file
store.

There are no specs currently for the `#directory` and `#filepath`
methods. This due to a `if Rails.env.test?` conditional in the model.
These methods won't be used when a install switches to `ActiveStorage`
and so they can be deprecated.