Refactor conditional reindexing of InfoRequest
`reindex_some_request_events` is a misleading name, in that we're
reindexing _all_ events, but only if a subset of attributes have
changed.

This commit makes this behaviour clearer by moving the condition to its
own method (`reindexable_attribute_changed?`) and using a callback
condition to decide whether to trigger `reindex_request_events`.