Prevent empty embargo_duration
Some links to batch requests were not working because the
`InfoRequestBatch#embargo_duration` was an empty `String`, but our
`Ability` checks were expecting only `nil`. (h/t to @gbp for figuring
this out).

The issue is that when the batch is created without an embargo, the form
submission submits an empty string rather than `nil` for the
`embargo_duration` param.

This commit introduces a targeted use of `strip_attributes` to fix this
specific problem. We should mirror the `strip_attributes` calls around
all the models that result in the creation of an `InfoRequest`.

Also fixes the same problem for the currently unused
`ProAccount#default_embargo_duration` to prevent future issues related
to empty `embargo_duration` values.

Fixes https://github.com/mysociety/alaveteli/issues/5584