Localise salutation stripping
Use the translated salutation in favour of the hard-coded English
salutation.

The one downside here is that we wouldn't catch "Dear Person Name," now,
but that shouldn't be a huge problem since all the auto-generated
salutations use the public body name.

Need to check for `public_body` now that we're generating the
salutation.

Fixes https://github.com/mysociety/alaveteli/issues/2748.