Fix the load-mail-server-logs script
The script cd into the `script` directory which means the new
`bin/config` command isn't found when retriving the MTA_LOG_PATH config.

This change replicates `script/mailin so it cd into the root directory
and the config can now be returned.

Also adds specs for this script to prevent other errors in the future.

Fixes #6639