Reduce prominence of User#about_me
We get lots of spam signups which mainly populate the `about_me` with
links to some form of junk site. Many of these spam signups originate
through another spam user profile, so presumably they're searching for
particular terms in the spam text in order to find sites that can be
targeted.

This commit aims to reduce the attractiveness of Alaveteli to spammers
by hiding the `about_me` text from the wider web, except for when the
user has been manually confirmed as genuine.

We'll show the text to logged in users, unless the user being viewed has
been banned or closed as before.

Fixes https://github.com/mysociety/alaveteli/issues/6627.