Add limit to mirrored blobs to be promoted
When running the `storage:mirror` and the `storage:promote` tasks we
loop through all raw emails in the configured `ActiveStorage` mirror
service.

While it's fine for all raw emails to be mirrored, we want to limit the
promotion for a week so the primary service (which normally should be a
local disk) is used when caching the response initially.

The theory is this will reduce the number of downloads from secondary
service (normally a cloud service) as the response should be parsed
normally within 7 days of being received.