Add external request handling for event digest
If a external request, which doesn't have an associated user record, is
updated and matches a TrackThing then an exception will be raise causing
the alert track daemon to fail.

This change check to see if the request is external and substitute "An
anonymous user" in place of the user name. This is the same string which
appears on the request#show view.