Add `after_initialize` block to set locales
When upgrading to Rails 6 [1] this was moved from an standard
initializer to `before_configuration` block as this was required for a
new deprecation warning which happened to includes an I18n translation,
requiring the default locale to be configured before the default Rails
initializers ran.

Downside to this is it broke the fallback options for locales with
underscores (EG `en_RW`). This change required to get these locales
working correctly again when loading content from Globalize translation
tables.

Unfortunately we can't easily test this ans the test environment is run
with `config.cache_classes = true`, this is disabled in development,
hence the conditional added to the duplicate code remaining in the
`before_configuration` block.

Fixes #5382

[1] See https://github.com/mysociety/alaveteli/commit/02ea505