Add "Community" link to main navigation
Make it clearer that there's a whole Alaveteli community. The footer and
a main section on the homepage have a "Contact" link that also links to
`/community`, so I don't think we're losing anything by making this
change. It's pretty natural to scroll to the footer to find the contact
details IME.

Fixes https://github.com/mysociety/alaveteli/issues/1897