Improve admin layout document titles
Prefixing the site name makes it difficult to differentiate between
several admin tabs. This puts the page title first and then adds the
"Admin" on the end.

We lose a _little_ here in that it's less obvious a given tab is an
_admin_ page rather than a public page, but an easy indicator is whether
the record is prefixed with the model class name (e.g. "User – Bob
Smith"). Public pages would just be "Bob Smith".