Add extra one-click user ban reasons
Adds a couple of extra one-click ban reasons to the one-click "Ban for
spamming" admin button.

Ideally we'd move these to canned reasons for the `User#ban_text` form
field, but there's a bit of complication with that being a `<text_area>`
rather than an `<input>`. For now this adds some common reasons that
help with common cases.

A simplified version of https://github.com/mysociety/alaveteli/issues/6699.