Minor styling of user admin button toolbar
* Make "Edit" a standard button – it's not really a "primary" action;
  just an occasional action.
* Group "Edit" and "Public page" as these are closely related (i.e
  seeing the record in slightly different ways).
* Bring the "login as" button into the main `btn-toolbar`.
* Use a `link_to` for "login as" to reduce template complexity.
* Add a `btn-toolbar` around the "ban for" `btn-group` for alignment.

Makes the page feel less messy which makes it more pleasant to be in
when you're working with the interface.

I did wonder about the button class of "Login as". We seem to use
`btn-info` for creating secondary associated content (like Censor
Rules), which is not what we're doing here. However, `btn-warning` felt
too strong, and a plain `btn` didn't feel differentiated enough. For now
it's fine as it is.