Add prominence_icon to admin user listings
Allows an at-a-glance look at whether a User is banned, inactive or
otherwise in a compact method. When banned a user gets “Account
Suspended” prepended to their name, and in some cases an orange “banned”
label, but in other more information-dense listings we either don’t show
this or it’s hard to parse out.

This makes it easy to tie banned users to hidden content or hidden
content to unbanned users.