Record IP addresses associated with User signins
If a retention period is configured, create a record associated with the
user of the signin and the IP address used to do so.

Uses Postgres' inet so that we can handle both IPv4 and IPv6 addresses.
Fake ipv6 taken from Faker example [1].

A rake/cron task is provided to purge records older than the retention
period.

A version of https://github.com/mysociety/alaveteli/issues/2691

[1] https://github.com/faker-ruby/faker/blob/a00b4c6388c69fdafb7e12cbacdb6d8a9927fba5/lib/faker/default/internet.rb#L423