Update extraction of multipart attachments
So we extract and save as two distinct steps. Do not create or save
these attachments instead initialise the instance assign the attributes
to be saved later. This will allow us to save all attachments in a
single call to `#save!`.

For any attachment which doesn't exist we can mark as needing to be
destroyed on save.