Fix reading attachments inside transaction
When parsing incoming messages we were reloading the `foi_attactments`
association. This removed access to the attachment body stored as an
instance variable.

Reloading of the attachments breaks the Xapian update index task for any
unparsed emails. This been an issue since migrating `FoiAttachment` to
use `ActiveStorage`.

This is because `ActiveStorage` uploads files after the transaction has
been committed.

We should be able to rely on the instance variable `@cached_body` but
reloading the association means the Xapian update index task has a
different instance of the attachment without the `@cached_body` variable
set and we can't yet read the file from `ActiveStorage` until the
transaction is committed.