Improve InfoRequest#external_user_name test data
"External User" looks quite generic, as if it could be automatically
generated. This changes makes it clear that the value returned is an
individual's name.

"Bob Smith" is an existing fixture user. Using the same name makes it
difficult to quickly understand whether you should be expecting an
internal or external request, so rename this to create a stronger
differentiation.