Reduce complexity of AdminRequestController#destroy
The rendering of the `User#email` after destroying the request doesn't
add much value. Destroying a request doesn't destroy the user, so we can
always look up the user to get their email address. There's a slight
downside in that if you destroy a request without first having the
user's admin page open you may not be able to trace back to the user,
but this seems like a very minor issue given the flash message would be
removed after a page reload.

It was introduced in 010dbab8553 and then tweaked in 96e0815eeda to
account for external users.