Remove file extensions from render partials
When passing partials through the flash to be rendered on the next page
load or rendering partials inline we used to include the file extension.

This has been deprecated in Rails, such as:
  `Rendering actions with '.' in the name is deprecated`

When the default `html` extension isn't used, we now have to pass the
`formats` argument with the file format instead of the extension.

Fixes #6892