Update query order/reorder syntax
Remove string based ordering and replace with symbols/hashes. This has
the benefit of allowing `ActiveRecord` to automatically adding the
correct database table prefix before each column. This is important when
using joins as sometimes there will be column name clashes. In Rails 7
this appears to be an issue with some of the `has_many` condition procs.

For ordering on joins tables we can't use symbols/hashes, so instead of:
  `Klass.joins(:join_klass).order('join_klasses.column')`
we can use the:
  `Klass.joins(:join_klass).merge(JoinKlass.order(:column))`

When doing this I needed to switch from using `includes` which adds the
join tables columns to query select, to `joins` and `preload`, so an
single extra database query is run.