Update PublicBody#expire_requests delegation
This method is called when a `PublicBody#name` is changed as the
`#url_name` will also get updated.

Delegating to `InfoRequest#expire` isn't ideal as this method will drop
database cache columns for any incoming messages.

This means all messages received from a given `PublicBody` will be
reparsed when `#name`/`#url_name` is changed. This very costly and
something admins should be able to do without impacting the site.