Update queries loading headings and categories
Reduce the number of queries required to load all public body headings
and categories.

This executes one query for the linked table to get the correct ordering
of headings and categories.

We also preload the all headings and categories (and their translations)
resulting just 5 queries being executed.

Before this change we had three separate n+1 queries after loading the
headings, we loaded:
1. translation instance for each heading
2. categories for each heading
3. translations for each category