Make it clearer that users are making info requests
Sometimes users use Alaveteli to make general or unsuitable requests. It
is likely they are less digitally literate and believe the authority
page to be an official or general contact form.

This tweak makes it clearer that users will be making an info request.

Part of https://github.com/mysociety/alaveteli/issues/7102