Refactor admin tag helper
With the introduction of an admin tag controller we will be able to link
to an admin view. This refactors the existing helpers for rendering tags
so that all tags rendered in the admin are link.

This removes the custom `search_target` for public body tags, linking to
the public page with handled via the upcoming admin `tags#show` view.

At present all links are targeting an anchor on the current page. This
will be updated when we add an admin `tags#show` action.