Add admin title class method to models
Some of our models names can't be derived from the class name. This adds
an `#admin_title` class method to these models so we can dynamically
referred to within the admin interface.

In most Rails apps this would be done via a `config/locales/en.yml` file
but as we're using gettext we don't have this option at present. I think
this approach is better than adding a second translation layer.