Allow render_tag helper to accept a string tag
Instead of needing a `HasTagString::HasTagStringTag` instance this
commit changes the helper to also accept a text string.

Will correctly link and style the string tag with name/value to the tag
admin controller.

As this doesn't include the `model_type` so the tag controller will
assume `PublicBody` which is fine and a good assumption to make.