Allow admins to destroy post redirects
We often spot user accounts with unconfirmed requests that don't meet
our house rules. The only way to prevent a user going on to confirm the
request is by destroying its associated `PostRedirect`, which until this
commit could not be performed through the admin UI.

This action currently only handles post redirects with an associated
`User`. We do create post redirects _without_ associated an associated
`User` record, but those are not displayed in the admin UI anywhere, so
I don't think that's an immediate issue.

Fixes https://github.com/mysociety/alaveteli/issues/3500