Rationalise links in incoming & outgoing edit view
When administering messages it’s quite annoying to have different ways –
or no way at all – of getting back to the public content or the admin
page for the request.

This commit rationalises the two admin views so that there’s the ability
and consistency of being able to navigate to the right content.

The one thing that’s been dropped is the link to “List all”, but that’s
easily available from the main admin nav bar.

The use of both_links also adds richer information in the form of
showing the current prominence for both request and message.