Update `all-authorities.csv` generation
Generate this file once a night and serve cached file in response to
hits to this endpoint.

Saves regenerating for each request as this can take some time and block
passenger processes until complete.

Moves the generated file from `/cache/zips/...` to just `./cache` as
this the file isn't zipped.