Use both_links on message listings
We have a `both_links` for outgoing and incoming messages so we can use
it here for more conventional behaviour.

Adds the benefit of being able to quickly understand the prominence of
various messages on a thread.

Mutes the more incidental metadata about the messages.

The rendered `both_links` adds quite a bit of duplication of the request
name in this context, but I think it's good to keep `both_links`
flexible enough to be used in any context (e.g. a page of its own). Just
dropping it down to `#outgoing-123` would make it pretty meaningless to
view on e.g. a page of failed-to-send outgoing messages.