Prevent users from accessing embargoed content
Admins are able to hide requests which are `not_foi` or `vexatious`. The
application will automatically generate a explanation to the user why
their request is hidden.

The action which generates the explanation message has no authentication
so crafting a URL could expose the requester's name and subject of the
request regardless of if the current user can read the request.

This adds a `before_action` block, similar to what we recently added to
some admin controllers [1], to ensure the user can `read` the request.

[1] https://github.com/mysociety/alaveteli/pull/7123