Add request prominence counts in admin UI
Make it easy to quickly see the distribution of prominence states for
requests on body and user admin pages.

The use case for this was wanting to know if a request hidden due to a
data breach was the first for the authority in question or yet another
in a long history of breaches.

We do try to minimise reduced prominence to specific messages or even
attachments after #1005 [1], so only showing request prominence counts
doesn't tell the whole story, but it's a start.

[1] https://github.com/mysociety/alaveteli/issues/1005