Update Pro classification controller auth
Former pro users still need to be able to update their request statuses.

This change removes the `before_action` callback to authenticate the pro
users. This means all requesters can use this endpoint to classify their
request but in reality only current and former Pro users will as this is
only linked to from the Pro UI.

Fixes #7370