Remove scopes from abilities
These are causing issues in production when starting the application.

Seems to be due to the large number of requests in WDTK production and
the scopes and/or CanCanCan not being able to correctly handle this.

Reverting back to using blocks but keeping the more verbose `can :_read`
abilities from bdf93b85336cbbcde3a9ab3d391c34d0dfccdb32