Remove User#otp_secret_key from admin columns
The OTP secret key for a user is visible to admins, which means an admin
could configure an OTP app with someone else's secret, if we were using
authenticator apps.

Our current OTP system requires the user to store an OTP that's been
generated, and is only changed on use.

While impractical, it may be possible for a malicious admin to inspect
the source code of active_model_otp to figure out how to generate the
user's current OTP using information available in the admin interface.
Removing the `otp_secret_key` attribute from the admin columns
eradicates this risk. A malicious actor would then need console or
database access.

Fixes https://github.com/mysociety/alaveteli/issues/2923.