Remove custom will_paginate CSS class
We don't have a `.paginator` class defined, so remove it.

This also acts as an improvement, since Bootstrap's pagination class is
`.pagination`, which is also the default applied by `will_paginate`. As
such, we get some better spacing built in.

Bootstrap's styling doesn't apply though, since it requires items to be
contained within a `<ul>`.