Render User#about_me in the admin list of users
To check whether a user is a spammer, admins have to open the user page,
review the text, ban the user (if spam), then go back to the results
list to look at the next one.

Having "About me" in the results list means the admin doesn't need to
leave the list, as we already have the "Ban" button in the unfolded
accordion.

Fixes https://github.com/mysociety/alaveteli/issues/5599