Add InfoRequestEvent#event_type scopes
Dynamically add a scope per event type to make querying the model
easier.

The specs check the generated SQL because setting message expectations
didn't work. For example:

    before { expect(described_class).to receive(:where) }
    it { described_class.public_send(event_type) }

In this example I was expecting the `where` to be called, but it never
got called. Perhaps Rails is doing something different with scopes? I
tried adding a `#to_a` call to the example to force the query to
actually run, but nope.

I think asserting that the generated SQL gives us enough confidence for
what this is doing, so went with that.