Privatise InfoRequestEvent methods
With the exception of `get_clipped_response_efficiently` and
`must_be_valid_state`, all the newly-private methods are only used as
helpers for the `acts_as_xapian` configuration, so can be made private
to reduce the public surface area of the object.

`get_clipped_response_efficiently` is only called in `search_text_main`,
which again is a helper for the `acts_as_xapian` configuration (though
is also called elsewhere to render search result snippets).

`must_be_valid_state` is only called via a `validate` call so there's no
need for it to be public.

`variety` is a slightly special case in that it's called on sibling
events via `#latest_variety`, so it's made `protected` instead of
`private`.

I thought it worth preserving the specs to ensure we're populating the
search index with correct data, though it does mean using `#send` to
call the methods. I imagine we can improve this down the line. It seems
more valuable to add clarity over internal vs external methods at this
point.