Allow ex-pro users to followup to private requests
Ex-pro users should still be able to manage requests even if they're
still private; they just can't extend privacy periods or create new
ones.

We don't need to check that the `current_user` is a Pro here; we first
search the user's requests, irrespective of privacy state, and assign to
the instance variable. If that isn't found – which means the user is
trying to follow up to someone else's request – we only search public
requests in order to display the notice to sign in to the request
author's account.

Fixes https://github.com/mysociety/alaveteli/issues/7456.