Output CSV for users:count_per_domain
CSV is machine parsable so it's much more useful to be able to pipe this
to a file and use common tools to filter and extract information from
the output. The previous format wasn't much more readable than CSV.

Also fixes some minor code style issues and improves the task
description.