Add ability to paginate through batch requests
Add "Next" and "Prev" pagination links to the sidebar of info requests
that belong to a batch request.

This allows users to navigate through requests in the batch without
having to click "in and out" via the batch page.

To minimise implementation complexity the pagination doesn't have a
beginning and end. We just cycle back around to the first request in the
batch when we hit the last one; same for paginating in reverse. If we
didn't do this we'd have to have conditionals for disabling the
particular navigation option when we hit an end.