Add inline help around censor rule ordering
It's not obvious that censor rules are applied in order, and each
applies to the text after it's been modified by the previous rule, so
some combinations of rules can do unexpected things.

For example, we often want to redact "firstname lastname" and
"firstname". If the rule for "firstname" is applied first, then the
"firstname lastname" rule will not apply, because the text cannot be
matched due to the "firstname" redaction.

Fixes https://github.com/mysociety/alaveteli/issues/7495.