Remove next Gemfile
This was used to help upgrade Rails versions but since we're now on the
latest version this isn't as useful.

In fact it is causing headaches when dealing with dependency upgrades as
we have to manually keep the two lockfiles in sync.

We attempted to use `bootboot` and while it helped a bit the Dependabot
PRs still didn't stage the next lockfile changes.

I had though keeping the next Gemfile would be good for helping with
Ruby and other larger dependency upgrades but we can handle these in
isolation is separate PRs.