Reduce default visibility of sign-in tracking
IP addresses are quite identifying, and in WhatDoTheyKnow's privacy
policy [1] we say that we don't ordinarily look at them while
administering the site.

Making it an active decision by an admin to reveal these preserves this
position of default privacy, but makes it much easier to investigate
issues without developer involvement.

[1] https://www.whatdotheyknow.com/help/privacy#our_logging