Make it clearer who a request is being sent from
This change attempts to reduce two types of support case.

1. Users signing off with a pseudonym and then being confused that their
   account name was sent to the authority.
2. Users not realising their user name will be published and sent to the
   authority.

This adds the user name to the request header to make it clear that
their name will be used. Where we don’t have a name because a user is
drafting a request before they’re signed in we just render an ellipsis.

The latter case – and point 2 – could be improved with some inline help
(“We’ll add your name when you register an account or sign in”, and
“Your name will be sent to the authority and published”) but we
don’t have existing styling for that so it can be a later addition.

A step towards https://github.com/mysociety/alaveteli/issues/6154.