Add comment creation interval rate limit
Calculates whether too many comments are being created too quickly.

We don't want to allow a user to sign up and smash through their daily
limit in 5 minutes, then create a new account to do the same, and so on.

This ensures that the allocation is used sensibly. It hopefully won't
affect the majority of genuine users.

Inverts the logic of `CommentController#reject_if_user_banned` now that
the daily limit and rate limit are checked in `User#can_make_comments?`.
Now, the user being banned is the edge case. This doesn't quite make
sense for inactive users, but that wasn't handled in
`User#can_fail_html` either.