Clarify logic checking whether user can comment
A previous version [1] of this check was defined as:

    if authenticated? && !authenticated_user.can_make_comments?
      # render the correct view for not being able to comment

The condition was moved to a guard clause [2] to improve code style, but
the move to using `unless` makes it less human readable.

The way I’d describe the clause naturally is:

> Skip the check if there isn’t a signed in user, or if the signed in
  user can make comments.

This commit tweaks the implementation to match this more natural
phrasing.


[1] 9c28c412
[2] e895b773