Refactor attachment controller specs
Combined lets and variables for `info_request`, `incoming_message` and
`attachment` so that they are declared once per context.

This is so we can make changes easier to all specs. For example, this
will be useful for, adding a `masked` trait to the `FoiAttachment`
factory. In coming commits we will be updating the controller flow to
work differently if an attachment isn't maksed, adding this trait will
keep the original behaviour and mean we won't need to update specs on
mass.

Also minor refactoring to:
- Use `show` and `show_as_html` action helpers throughout
- Move specs into a single describe block for each action
- Remove blank lines inside describe and context blocks
- Don't wrap describe or it labels
- Fix linting; indentation, string quotes, etc...