feat: add censor rules JS and use previous comment feature in admin_censor_rule form
This change introduces a new JavaScript file, censor-rules.js, to handle the functionality for the admin censor rules section. Specifically, it adds a "Use previous comment" feature to the form for editing censor rules. The change has been made by:

Including the new censor-rules.js file in admin.js to make it available for use in the admin section of js files.

Creating the censor-rules.js file, where a click event listener is added to the "Use previous comment" link. When the link is clicked, the event handler retrieves the last edit comment from the associated text area's data attribute and sets the text area's value to the comment at the last edit.

Updating the _form.html.erb partial for the censor rule form. In the partial, a data attribute is added to the last_edit_comment text area to store the last_edit_comment value from the censor_rule object.

The purpose of this change is to streamline the editing process for admin users. By providing a "Use previous comment" option, admins can easily reuse the last edit comment as the reason for the current update, saving time and effort when making similar or related changes to censor rules.