Remove file_watcher from config
In Rails 7, the file_watcher variable was removed from the config
template (as well as the listen gem).

This change is now reflected here, which makes development in a
docker environment easier, because it means that the container
doesn't need to be restarted for changes to files to reflect
in the local website in the browser.