Use friendly_id history module
Needs to override some internal friendly_id methods to use the
non-history module version.

These methods relate to the generation and searching for slugs as this
will currently fail due to history module expecting this information to
be in `FriendlyID::Slug` instances but until we migrate the existing
records to this model is doesn't exist.

We will add a `temporary` rake task to migrate slugs.