refactor: email helper methods are more verbose
receive_incoming_mail and gsub_addresses methods (in email_helpers.rb)
now both take optional named parameters for various email headers, with
the intention that tests regarding email headers are slightly clearer to
read, and easier to write.

The scope of this commit extends to changing the specs that use either
of these methods to conform to the new parameters, so this isn't a
breaking change.