Drop fragment cache for batch category picker
With output cached the category picker doesn't function correctly due to
the wrong draft ID being cached.

We could add the draft to the cache key but I think performance isn't a
concern here anymore since some other improvements we made for #6862

Ultimately the categories UI shouldn't include the draft ID (or detect
is a body has been added or not), the JS should handle rendering and
handle the "Add" events.

I think it is how it is due to us trying to keep the interface working
for non-JS users but this isn't worth the hassle especially as Pro users
need JS to sign up via Stripe.

Fixes #7813